Move credentials class to commons
