+++ whois-api/src/main/java/net/ripe/db/whois/api/UpdatesParser.java
@@ -4,17 +4,17 @@ import com.google.common.base.Splitter;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import net.ripe.db.whois.common.Latin1Conversion;
-import net.ripe.db.whois.update.domain.OAuthCredential;
-import net.ripe.db.whois.update.domain.ClientCertificateCredential;
+import net.ripe.db.whois.common.Credentials.OAuthCredential;
+import net.ripe.db.whois.common.Credentials.ClientCertificateCredential;
 import net.ripe.db.whois.update.domain.ContentWithCredentials;
-import net.ripe.db.whois.update.domain.Credential;
+import net.ripe.db.whois.common.Credentials.Credential;
 import net.ripe.db.whois.update.domain.Credentials;
 import net.ripe.db.whois.update.domain.Operation;
-import net.ripe.db.whois.update.domain.OverrideCredential;
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
 import net.ripe.db.whois.update.domain.Paragraph;
-import net.ripe.db.whois.update.domain.PasswordCredential;
+import net.ripe.db.whois.common.Credentials.PasswordCredential;
 import net.ripe.db.whois.update.domain.PgpCredential;
-import net.ripe.db.whois.update.domain.SsoCredential;
+import net.ripe.db.whois.common.Credentials.SsoCredential;
 import net.ripe.db.whois.update.domain.Update;
 import net.ripe.db.whois.update.domain.UpdateContext;
 import net.ripe.db.whois.update.domain.UpdateMessages;
