+++ whois-api/src/main/java/net/ripe/db/whois/api/mail/dequeue/MessageDequeue.java
@@ -13,7 +13,7 @@ import net.ripe.db.whois.common.DateTimeProvider;
 import net.ripe.db.whois.common.MaintenanceMode;
 import net.ripe.db.whois.common.Messages;
 import net.ripe.db.whois.update.domain.DequeueStatus;
-import net.ripe.db.whois.update.domain.PasswordCredential;
+import net.ripe.db.whois.common.Credentials.PasswordCredential;
 import net.ripe.db.whois.update.domain.Update;
 import net.ripe.db.whois.update.domain.UpdateContext;
 import net.ripe.db.whois.update.domain.UpdateMessages;
