+++ whois-api/src/test/java/net/ripe/db/whois/api/UpdatesParserTest.java
@@ -3,11 +3,11 @@ package net.ripe.db.whois.api;
 import com.google.common.collect.Lists;
 import net.ripe.db.whois.common.rpsl.RpslObject;
 import net.ripe.db.whois.update.domain.ContentWithCredentials;
-import net.ripe.db.whois.update.domain.Credential;
+import net.ripe.db.whois.common.Credentials.Credential;
 import net.ripe.db.whois.update.domain.Operation;
-import net.ripe.db.whois.update.domain.OverrideCredential;
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
 import net.ripe.db.whois.update.domain.Paragraph;
-import net.ripe.db.whois.update.domain.PasswordCredential;
+import net.ripe.db.whois.common.Credentials.PasswordCredential;
 import net.ripe.db.whois.update.domain.PgpCredential;
 import net.ripe.db.whois.update.domain.Update;
 import net.ripe.db.whois.update.domain.UpdateContext;
