+++ whois-commons/src/main/java/net/ripe/db/whois/common/Credentials/Credential.java
@@ -1,0 +1,5 @@
+package net.ripe.db.whois.common.Credentials;
+
+public interface Credential {
+
+}
