+++ whois-update/src/main/java/net/ripe/db/whois/update/authentication/Authenticator.java
@@ -15,8 +15,8 @@ import net.ripe.db.whois.common.rpsl.RpslObject;
 import net.ripe.db.whois.update.authentication.strategy.AuthenticationFailedException;
 import net.ripe.db.whois.update.authentication.strategy.AuthenticationStrategy;
 import net.ripe.db.whois.update.domain.Origin;
-import net.ripe.db.whois.update.domain.OverrideCredential;
-import net.ripe.db.whois.update.domain.PasswordCredential;
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
+import net.ripe.db.whois.common.Credentials.PasswordCredential;
 import net.ripe.db.whois.update.domain.PreparedUpdate;
 import net.ripe.db.whois.update.domain.UpdateContext;
 import net.ripe.db.whois.update.domain.UpdateMessages;
