+++ whois-update/src/main/java/net/ripe/db/whois/update/authentication/credential/OAuthCredentialValidator.java
@@ -5,9 +5,9 @@ import net.ripe.db.whois.common.Messages;
 import net.ripe.db.whois.common.oauth.APIKeySession;
 import net.ripe.db.whois.common.oauth.OAuthSession;
 import net.ripe.db.whois.common.rpsl.RpslObject;
-import net.ripe.db.whois.update.domain.OAuthCredential;
+import net.ripe.db.whois.common.Credentials.OAuthCredential;
 import net.ripe.db.whois.update.domain.PreparedUpdate;
-import net.ripe.db.whois.update.domain.SsoCredential;
+import net.ripe.db.whois.common.Credentials.SsoCredential;
 import net.ripe.db.whois.update.domain.Update;
 import net.ripe.db.whois.update.domain.UpdateContext;
 import net.ripe.db.whois.update.log.LoggerContext;
