+++ whois-update/src/main/java/net/ripe/db/whois/update/authentication/credential/SsoCredentialValidator.java
@@ -2,7 +2,7 @@ package net.ripe.db.whois.update.authentication.credential;
 
 import net.ripe.db.whois.common.rpsl.RpslObject;
 import net.ripe.db.whois.update.domain.PreparedUpdate;
-import net.ripe.db.whois.update.domain.SsoCredential;
+import net.ripe.db.whois.common.Credentials.SsoCredential;
 import net.ripe.db.whois.update.domain.Update;
 import net.ripe.db.whois.update.domain.UpdateContext;
 import net.ripe.db.whois.update.log.LoggerContext;
