--- whois-update/src/main/java/net/ripe/db/whois/update/domain/Credential.java
@@ -1,5 +1,0 @@
-package net.ripe.db.whois.update.domain;
-
-public interface Credential {
-
-}
