+++ whois-update/src/main/java/net/ripe/db/whois/update/domain/OverrideOptions.java
@@ -1,6 +1,7 @@
 package net.ripe.db.whois.update.domain;
 
 import com.google.common.base.Splitter;
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
 
 import javax.annotation.concurrent.Immutable;
 import java.util.Set;
