+++ whois-update/src/main/java/net/ripe/db/whois/update/domain/Update.java
@@ -1,5 +1,6 @@
 package net.ripe.db.whois.update.domain;
 
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
 import net.ripe.db.whois.common.rpsl.ObjectType;
 import net.ripe.db.whois.common.rpsl.RpslObject;
 import org.apache.commons.lang.Validate;
