+++ whois-update/src/main/java/net/ripe/db/whois/update/domain/UpdateContext.java
@@ -3,6 +3,7 @@ package net.ripe.db.whois.update.domain;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
 import net.ripe.db.whois.common.Message;
 import net.ripe.db.whois.common.Messages;
 import net.ripe.db.whois.common.oauth.OAuthSession;
