+++ whois-update/src/test/java/net/ripe/db/whois/update/domain/ContentWithCredentialsTest.java
@@ -1,6 +1,7 @@
 package net.ripe.db.whois.update.domain;
 
 import com.google.common.collect.Lists;
+import net.ripe.db.whois.common.Credentials.Credential;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
