+++ whois-update/src/test/java/net/ripe/db/whois/update/domain/OverrideOptionsTest.java
@@ -1,5 +1,7 @@
 package net.ripe.db.whois.update.domain;
 
+import net.ripe.db.whois.common.Credentials.Credential;
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
 import org.junit.jupiter.api.Test;
 
 import org.mockito.Mock;
