+++ whois-update/src/test/java/net/ripe/db/whois/update/domain/UpdateTest.java
@@ -2,6 +2,8 @@ package net.ripe.db.whois.update.domain;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import net.ripe.db.whois.common.Credentials.OverrideCredential;
+import net.ripe.db.whois.common.Credentials.PasswordCredential;
 import net.ripe.db.whois.common.rpsl.RpslObject;
 import org.junit.jupiter.api.Test;
