+++ whois-update/src/test/java/net/ripe/db/whois/update/log/UpdateLogTest.java
@@ -7,8 +7,8 @@ import net.ripe.db.whois.common.rpsl.ObjectMessages;
 import net.ripe.db.whois.common.rpsl.RpslObject;
 import net.ripe.db.whois.update.domain.Action;
 import net.ripe.db.whois.update.domain.Credentials;
-import net.ripe.db.whois.update.domain.OAuthCredential;
-import net.ripe.db.whois.update.domain.PasswordCredential;
+import net.ripe.db.whois.common.Credentials.OAuthCredential;
+import net.ripe.db.whois.common.Credentials.PasswordCredential;
 import net.ripe.db.whois.update.domain.Update;
 import net.ripe.db.whois.update.domain.UpdateContext;
 import net.ripe.db.whois.update.domain.UpdateRequest;
